+++ trunk/Source/Clients/Health/GDataQueryGoogleHealth.h
@@ -29,16 +29,16 @@
 - (void)setIsGrouped:(BOOL)flag;
 - (BOOL)isGrouped;
 
-- (void)setMaxResultsGroup:(int)val;
-- (int)maxResultsGroup;
+- (void)setMaxResultsGroup:(NSInteger)val;
+- (NSInteger)maxResultsGroup;
 
-- (void)setMaxResultsInGroup:(int)val;
-- (int)maxResultsInGroup;
+- (void)setMaxResultsInGroup:(NSInteger)val;
+- (NSInteger)maxResultsInGroup;
 
-- (void)setStartIndexGroup:(int)val;
-- (int)startIndexGroup;
+- (void)setStartIndexGroup:(NSInteger)val;
+- (NSInteger)startIndexGroup;
 
-- (void)setStartIndexInGroup:(int)va;
-- (int)startIndexInGroup;
+- (void)setStartIndexInGroup:(NSInteger)va;
+- (NSInteger)startIndexInGroup;
 
 @end
